An example of PiePieSpider.
This spider is an example of using PiePieSpider. And this spider is a non-extensive spider, that means there are no items new to the spider queue after you initialize the queue. People can use this spider to get query result of Cnki. To get query result, you are suggested to submit a formdata. The details are in the files CnkiQueryTemplate.py and CnkiNetWalker.py.